requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
pg8000==1.30.4
python-dotenv==1.0.0
Flask-WTF==1.1.1